[memory] Restruct memory module
- Split memory.hpp into separate headers
